divided into two classes, move Rakefile to lib/tasks
